requirements.txt
requests
pandas
plotly
openai
load_dotenv
kaleido
wordcloud 
matplotlib


